restruct ncurses functions
