+++ .gitignore
@@ -8,3 +8,4 @@ new_libft/obj/
 test_st.cor
 test_st.ror
 .idea/*
+src_vm/my_vm
